requirements.txt
Flask==2.0.1
flask-cors==4.0.0
gunicorn==21.2.0
pymongo[srv]==4.6.0
python-dotenv==1.0.0
certifi==2023.7.22
Werkzeug==2.0.1
dnspython==2.4.2
tenacity==8.2.3